Refactor remote actor handling to Remote_Actors class
Moved all remote actor management logic from Actors to a new Remote_Actors class. Updated references throughout the codebase to use Remote_Actors for fetching, updating, error handling, and inbox management of remote actors. Deprecated old methods in Actors and replaced their usage in collections, handlers, models, REST controllers, and templates. Added tests for Remote_Actors functionality.
