+++ includes/functions.php
@@ -11,6 +11,7 @@ use Activitypub\Activity\Activity;
 use Activitypub\Activity\Actor;
 use Activitypub\Activity\Base_Object;
 use Activitypub\Collection\Actors;
+use Activitypub\Collection\Remote_Actors;
 use Activitypub\Collection\Outbox;
 use Activitypub\Collection\Followers;
 use Activitypub\Collection\Following;
@@ -1561,7 +1562,7 @@ function follow( $remote_actor, $user_id ) {
 		return $remote_actor;
 	}
 
-	$remote_actor_post = Actors::fetch_remote_by_uri( $remote_actor );
+	$remote_actor_post = Remote_Actors::fetch_by_uri( $remote_actor );
 
 	if ( \is_wp_error( $remote_actor_post ) ) {
 		return $remote_actor_post;
@@ -1591,7 +1592,7 @@ function unfollow( $remote_actor, $user_id ) {
 		return $remote_actor;
 	}
 
-	$remote_actor_post = Actors::fetch_remote_by_uri( $remote_actor );
+	$remote_actor_post = Remote_Actors::fetch_by_uri( $remote_actor );
 
 	if ( \is_wp_error( $remote_actor_post ) ) {
 		return $remote_actor_post;
