+++ includes/handler/class-accept.php
@@ -8,7 +8,7 @@
 namespace Activitypub\Handler;
 
 use Activitypub\Notification;
-use Activitypub\Collection\Actors;
+use Activitypub\Collection\Remote_Actors;
 use Activitypub\Collection\Following;
 use Activitypub\Collection\Outbox;
 
@@ -54,7 +54,7 @@ class Accept {
 			return;
 		}
 
-		$actor_post = Actors::get_remote_by_uri( object_to_uri( $accept['object']['object'] ) );
+		$actor_post = Remote_Actors::get_by_uri( object_to_uri( $accept['object']['object'] ) );
 
 		if ( \is_wp_error( $actor_post ) ) {
 			return;
