+++ includes/handler/class-update.php
@@ -7,7 +7,7 @@
 
 namespace Activitypub\Handler;
 
-use Activitypub\Collection\Actors;
+use Activitypub\Collection\Remote_Actors;
 use Activitypub\Collection\Interactions;
 
 use function Activitypub\get_remote_metadata_by_actor;
@@ -113,6 +113,6 @@ class Update {
 			return;
 		}
 
-		Actors::upsert( $actor );
+		Remote_Actors::upsert( $actor );
 	}
 }
