+++ includes/model/class-follower.php
@@ -8,7 +8,7 @@
 namespace Activitypub\Model;
 
 use Activitypub\Activity\Actor;
-use Activitypub\Collection\Actors;
+use Activitypub\Collection\Remote_Actors;
 use Activitypub\Collection\Followers;
 
 use function Activitypub\extract_name_from_uri;
@@ -59,7 +59,7 @@ class Follower extends Actor {
 	 * @return mixed
 	 */
 	public function get_errors() {
-		return Actors::get_errors( $this->_id );
+		return Remote_Actors::get_errors( $this->_id );
 	}
 
 	/**
@@ -68,7 +68,7 @@ class Follower extends Actor {
 	 * @return bool True on success, false on failure.
 	 */
 	public function clear_errors() {
-		return Actors::clear_errors( $this->_id );
+		return Remote_Actors::clear_errors( $this->_id );
 	}
 
 	/**
@@ -106,7 +106,7 @@ class Follower extends Actor {
 	 * @return bool True on success, false on failure.
 	 */
 	public function reset_errors() {
-		return Actors::clear_errors( $this->_id );
+		return Remote_Actors::clear_errors( $this->_id );
 	}
 
 	/**
@@ -115,7 +115,7 @@ class Follower extends Actor {
 	 * @return int The number of errors.
 	 */
 	public function count_errors() {
-		return Actors::count_errors( $this->_id );
+		return Remote_Actors::count_errors( $this->_id );
 	}
 
 	/**
@@ -174,7 +174,7 @@ class Follower extends Actor {
 			return new \WP_Error( 'activitypub_invalid_follower', __( 'Invalid Follower', 'activitypub' ), array( 'status' => 400 ) );
 		}
 
-		$id = Actors::upsert( $this );
+		$id = Remote_Actors::upsert( $this );
 		if ( \is_wp_error( $id ) ) {
 			return $id;
 		}
