+++ includes/rest/class-following-controller.php
@@ -8,6 +8,7 @@
 namespace Activitypub\Rest;
 
 use Activitypub\Collection\Actors;
+use Activitypub\Collection\Remote_Actors;
 use Activitypub\Collection\Following;
 
 use function Activitypub\get_context;
@@ -113,7 +114,7 @@ class Following_Controller extends Actors_Controller {
 				\array_map(
 					function ( $item ) use ( $context ) {
 						if ( 'full' === $context ) {
-							$actor = Actors::get_actor( $item );
+							$actor = Remote_Actors::get_actor( $item );
 							if ( \is_wp_error( $actor ) ) {
 								return false;
 							}
