+++ includes/wp-admin/table/class-blocked-actors.php
@@ -8,6 +8,7 @@
 namespace Activitypub\WP_Admin\Table;
 
 use Activitypub\Collection\Actors;
+use Activitypub\Collection\Remote_Actors;
 use Activitypub\Collection\Blocked_Actors as Blocked_Actors_Collection;
 use Activitypub\Moderation;
 use Activitypub\Sanitize;
@@ -222,7 +223,7 @@ class Blocked_Actors extends \WP_List_Table {
 		);
 
 		foreach ( $blocked_actor_posts as $blocked_actor_post ) {
-			$actor = Actors::get_actor( $blocked_actor_post );
+			$actor = Remote_Actors::get_actor( $blocked_actor_post );
 			if ( \is_wp_error( $actor ) ) {
 				continue;
 			}
@@ -343,7 +344,7 @@ class Blocked_Actors extends \WP_List_Table {
 		$search = Webfinger::resolve( $search );
 
 		if ( ! \is_wp_error( $search ) && \filter_var( $search, FILTER_VALIDATE_URL ) ) {
-			$actor = Actors::fetch_remote_by_uri( $search );
+			$actor = Remote_Actors::fetch_by_uri( $search );
 			if ( ! \is_wp_error( $actor ) ) {
 				echo ' ';
 				\printf(
